+++ motion_planning/sbpl/src/planners/VI/viplanner.cpp
@@ -49,7 +49,7 @@ void VIPlanner::Initialize_vidata(CMDPSTATE* state)
 
 	vi_data->bestnextaction = NULL;
 	vi_data->iteration = 0;
-	vi_data->v = environment_->GetGoalHeuristic(state->StateID); 
+	vi_data->v = (float)environment_->GetGoalHeuristic(state->StateID); 
 
 
 }
@@ -295,7 +295,7 @@ void VIPlanner::backup(CMDPSTATE* state)
 
 	//set state values
 	((VIState*)state->PlannerSpecificData)->bestnextaction = minaction;
-	((VIState*)state->PlannerSpecificData)->v = minactionQ;
+	((VIState*)state->PlannerSpecificData)->v = (float)minactionQ;
 
 }
